requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.4
exceptiongroup==1.2.0
frozenlist==1.4.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-telegram-bot==13.13
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
StrEnum==0.4.15
telepot==12.7
tornado==6.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zope.interface==6.2
